[Fix] : 매치 ViewModel Domain Layer 구조 수정
