+++ app/src/main/java/com/link_up/matching_manager/data/repository/MatchRepositoryImpl.kt
@@ -1,25 +1,63 @@
 package com.link_up.matching_manager.data.repository
 
+import androidx.lifecycle.MutableLiveData
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
 import com.link_up.matching_manager.ui.match.MatchDataModel
 import com.link_up.matching_manager.domain.repository.MatchRepository
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.database.ServerValue
+import com.google.firebase.database.ValueEventListener
 import kotlinx.coroutines.tasks.await
 
 class MatchRepositoryImpl() : MatchRepository {
 
-    override suspend fun getList(database : FirebaseDatabase): List<MatchDataModel> {
+    override fun getList(database : FirebaseDatabase, originalList : MutableList<MatchDataModel>, list : MutableLiveData<List<MatchDataModel>>) {
         val matchRef = database.getReference("Match")
-        val items = arrayListOf<MatchDataModel>()
-        val snapshot = matchRef.get().await()
-        if (snapshot.exists()) {
-            for (matchSnapshot in snapshot.children) {
-                 matchSnapshot.getValue(MatchDataModel::class.java)?.let { matchData ->
-                     items.add(matchData)
+        matchRef.addListenerForSingleValueEvent(object : ValueEventListener {
+            override fun onDataChange(dataSnapshot: DataSnapshot) {
+                val dataList = mutableListOf<MatchDataModel>()
+
+                for (childSnapshot in dataSnapshot.children) {
+                    val matchData = childSnapshot.getValue(MatchDataModel::class.java)
+                    if (matchData != null) {
+                        dataList.add(matchData)
+                    }
                 }
+                val currentList = dataList
+                originalList.clear()
+                originalList.addAll(dataList)
+                list.value = currentList
             }
-        }
-        return items
+
+            override fun onCancelled(databaseError: DatabaseError) {
+                // 오류 처리
+            }
+        })
+    }
+
+    override fun autoGetList(
+        database: FirebaseDatabase,
+        list: MutableLiveData<List<MatchDataModel>>
+    ) {
+        val matchRef = database.getReference("Match")
+        matchRef.addValueEventListener(object : ValueEventListener {
+            override fun onDataChange(dataSnapshot: DataSnapshot) {
+                val dataList = mutableListOf<MatchDataModel>()
+
+                for (childSnapshot in dataSnapshot.children) {
+                    val matchData = childSnapshot.getValue(MatchDataModel::class.java)
+                    if (matchData != null) {
+                        dataList.add(matchData)
+                    }
+                }
+                list.value = dataList
+            }
+
+            override fun onCancelled(databaseError: DatabaseError) {
+                // 오류 처리
+            }
+        })
     }
 
     override suspend fun addData(data: MatchDataModel, database: FirebaseDatabase) {
