+++ app/src/main/java/com/link_up/matching_manager/ui/match/MatchViewModelFactory.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.link_up.matching_manager.data.repository.MatchRepositoryImpl
 import com.link_up.matching_manager.domain.usecase.match.MatchAddDataUseCase
+import com.link_up.matching_manager.domain.usecase.match.MatchAutoGetListUseCase
 import com.link_up.matching_manager.domain.usecase.match.MatchEditChatCountUseCase
 import com.link_up.matching_manager.domain.usecase.match.MatchEditViewCountUseCase
 import com.link_up.matching_manager.domain.usecase.match.MatchGetListUseCase
@@ -17,7 +18,8 @@ class MatchViewModelFactory : ViewModelProvider.Factory {
                 MatchAddDataUseCase(repository),
                 MatchGetListUseCase(repository),
                 MatchEditChatCountUseCase(repository),
-                MatchEditViewCountUseCase(repository)
+                MatchEditViewCountUseCase(repository),
+                MatchAutoGetListUseCase(repository)
             ) as T
         } else {
             throw IllegalArgumentException("Not found ViewModel class.")
